requirements.txt
Django==5.1.7
python-dotenv==1.0.1
requests==2.32.3
openai==1.65.5
 